requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu

